Tempo de carregamento melhorado, foi feita a solucao aonde carrega 30min para pelo menos mostrar em tela, enquanto em background e carregado o resto do exame

+++ PlotagemOpenGL/auxi/GlobVar.cs
@@ -181,7 +181,7 @@ namespace PlotagemOpenGL.auxi
         public static int[] grafSelected;
         public static int[] codSelected;
         public static short[,] matrizCompleta;
-        public static short[,] matrizCanal;
+        public static float[,] matrizCanal;
         public static int[] ponteiroI;
         public static int[] ponteiroF;
